less logic in NussknackerAppFactory
