+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/processreport/CountsReporterFactory.scala
@@ -1,0 +1,60 @@
+package pl.touk.nussknacker.ui.processreport
+
+import cats.effect.{IO, Resource}
+import com.typesafe.config.Config
+import com.typesafe.scalalogging.LazyLogging
+import pl.touk.nussknacker.engine.util.loader.ScalaServiceLoader
+import pl.touk.nussknacker.engine.util.multiplicity.{Empty, Many, Multiplicity, One}
+import pl.touk.nussknacker.processCounts.{CountsReporter, CountsReporterCreator}
+import pl.touk.nussknacker.processCounts.influxdb.InfluxCountsReporterCreator
+import pl.touk.nussknacker.ui.config.DesignerConfig
+import sttp.client3.SttpBackend
+
+import scala.concurrent.Future
+import scala.util.Try
+import scala.util.control.NonFatal
+
+object CountsReporterFactory extends LazyLogging {
+
+  def createCountsReporter(
+      designerConfig: DesignerConfig,
+      backend: SttpBackend[Future, Any]
+  ): Resource[IO, Option[CountsReporter[Future]]] = {
+    designerConfig.featureTogglesConfig.counts match {
+      case Some(config) => prepareCountsReporter(designerConfig.environment, config, backend)
+      case None         => Resource.pure[IO, None.type](None)
+    }
+  }
+
+  // by default, we use InfluxCountsReporterCreator
+  private def prepareCountsReporter(
+      env: String,
+      config: Config,
+      backend: SttpBackend[Future, Any]
+  ): Resource[IO, Option[CountsReporter[Future]]] = {
+    Resource
+      .make(
+        acquire = IO {
+          val configAtKey = config.atKey(CountsReporterCreator.reporterCreatorConfigPath)
+          val creator = Multiplicity(ScalaServiceLoader.load[CountsReporterCreator](getClass.getClassLoader)) match {
+            case One(cr) =>
+              cr
+            case Empty() =>
+              new InfluxCountsReporterCreator
+            case Many(many) =>
+              throw new IllegalArgumentException(s"Many CountsReporters found: ${many.mkString(", ")}")
+          }
+
+          Try(Option(creator.createReporter(env, configAtKey)(backend))).recover { case NonFatal(ex) =>
+            logger.warn(
+              s"Error while setting up counts mechanism: ${ex.getMessage}. Counts mechanism will be disabled."
+            )
+            None
+          }.get
+        }
+      )(
+        release = counter => IO(counter.foreach(_.close()))
+      )
+  }
+
+}
